refactor: eliminate debugConsoleTransport.ts
Problem:
- 'debugConsole' transport does not support "appendLine".
- debugConsoleTransport.ts is redundant with outputChannelTransport.ts

Solution:
Remove debugConsoleTransport.ts. Define 'debugConsole' using
outputChannelTransport.
