+++ src/test/shared/logger/sharedFileTransport.test.ts
@@ -8,7 +8,7 @@ import assert from 'assert'
 import { SharedFileTransport, flushIntervalMillis } from '../../../shared/logger/sharedFileTransport'
 import { FileSystemCommon } from '../../../srcShared/fs'
 import { stub, SinonStub } from 'sinon'
-import { MESSAGE } from '../../../shared/logger/debugConsoleTransport'
+import { MESSAGE } from '../../../shared/logger/consoleLogTransport'
 import { createTestFile } from '../../testUtil'
 import { readFileSync } from 'fs'
 import { sleep } from '../../../shared/utilities/timeoutUtils'
